requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
supabase==2.7.0
pydantic==2.9.2
pydantic-settings==2.6.0
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
openai==1.54.0
python-dotenv==1.0.1
websockets==14.0
aiofiles==24.1.0
httpx==0.27.0
redis==5.0.0
